screen2gif/requirements.txt
pip-requirements
pillow
imageio
pyautogui
numpy
PyQt5
mss
opencv-python
imageio-ffmpeg
pyperclip
pillow
imageio
pyautogui
numpy
